[ADD]- module/message, [FICS] - fics bag  in

+++ src/module/notification/notificant.service.ts
@@ -5,6 +5,13 @@ import { NotificationEntity } from '../database/entitis/notification.entitity';
 import { CustomLogger } from '../../helpers/logger/logger.service';
 import { Subject } from 'rxjs';
 
+interface SseEvent {
+  data: {
+    title: string;
+    body: string;
+  };
+}
+
 @Injectable()
 export class NotificationService {
   constructor(
@@ -14,18 +21,27 @@ export class NotificationService {
   ) {}
 
   // Храним подключения юзеров
-  private clients = new Map<number, Subject<any>>();
+  private clients = new Map<number, Set<Subject<SseEvent>>>();
 
   // Подключить юзера к SSE
-  subscribe(userId: number): Subject<any> {
-    const subject = new Subject<any>();
-    this.clients.set(userId, subject);
+  subscribe(userId: number): Subject<SseEvent> {
+    const subject = new Subject<SseEvent>();
+    if (!this.clients.has(userId)) {
+      this.clients.set(userId, new Set<Subject<SseEvent>>());
+    }
+    this.clients.get(userId)?.add(subject);
     return subject;
   }
 
   // Отключить юзера
-  unsubscribe(userId: number) {
-    this.clients.delete(userId);
+  unsubscribe(userId: number, subject: Subject<SseEvent>) {
+    const subjects = this.clients.get(userId);
+    if (!subjects) return;
+    subjects.delete(subject);
+    subject.complete(); // закрываем поток
+    if (subjects.size === 0) {
+      this.clients.delete(userId);
+    }
   }
 
   async sendNotification(
@@ -56,7 +72,7 @@ export class NotificationService {
     // Шлём SSE если юзер подключен
     const client = this.clients.get(userId);
     if (client) {
-      client.next({ data: { title, body } });
+      client.forEach((subject) => subject.next({ data: { title, body } }));
       this.logger.debug(`[SERVICE] SSE sent to userId: ${userId}`, refId);
     }
   }
@@ -74,10 +90,13 @@ export class NotificationService {
     }
   }
 
-  async markAsRead(id: number, refId: string) {
+  async markAsRead(id: number, userId: number, refId: string) {
     this.logger.debug(`[SERVICE] mark as read id: ${id}`, refId);
     try {
-      await this.notificationRepository.update(id, { isRead: true });
+      await this.notificationRepository.update(
+        { id, userId },
+        { isRead: true },
+      );
       this.logger.debug(`[SERVICE] mark as read SUCCESS id: ${id}`, refId);
     } catch (error) {
       this.logger.error(`[ERROR] mark as read: ${error}`, refId);
@@ -85,10 +104,10 @@ export class NotificationService {
     }
   }
 
-  async removeNotification(id: number, refId: string) {
+  async removeNotification(id: number, userId: number, refId: string) {
     this.logger.debug(`[SERVICE] remove notification id: ${id}`, refId);
     try {
-      await this.notificationRepository.delete(id);
+      await this.notificationRepository.delete({ id, userId });
       this.logger.debug(
         `[SERVICE] remove notification SUCCESS id: ${id}`,
         refId,
